requirements.txt
backoff==1.8.0
certifi==2020.12.5
chardet==4.0.0
ciso8601==2.1.3
idna==2.10
jsonschema==2.6.0
python-dateutil==2.8.1
pytz==2018.4
requests==2.25.1
simplejson==3.11.1
singer-python==5.12.1
six==1.15.0
urllib3==1.26.3
